Soporte de pesos con decimales, centrado de Kruskal eliminado. Proyecto listo para presentarse

+++ Angular/app.js
@@ -70,7 +70,7 @@ app.controller('mainCtrl', function($scope, $filter) {
         var ref=this;
         this.salida = salida;
         this.llegada = llegada;
-        this.costo = parseInt(costo, 10);
+        this.costo = parseFloat(costo, 10);
 
         $scope.$watch(
           function() {
@@ -272,10 +272,10 @@ app.controller('mainCtrl', function($scope, $filter) {
               //alert("CONEXION: " + JSON.stringify(conexion));
               if (!vecino.visto) {
                 //alert("No visto aún!!: " + JSON.stringify(vecino));
-                vecino.padre = parseInt(best_opt.id, 10);
-                vecino.costo = (parseInt(best_opt.costo, 10) + parseInt(conexion.costo, 10));
+                vecino.padre = parseFloat(best_opt.id, 10);
+                vecino.costo = (parseFloat(best_opt.costo, 10) + parseFloat(conexion.costo, 10));
                 //alert("NUEVO COSTO ES: " + vecino.costo);
-                //alert("Salida: " + salida.costo + "Costo Conexion: " + conexion.costo + " = " + (parseInt(salida.costo, 10) + parseInt(conexion.costo, 10)));
+                //alert("Salida: " + salida.costo + "Costo Conexion: " + conexion.costo + " = " + (parseFloat(salida.costo, 10) + parseFloat(conexion.costo, 10)));
                 vecino.iteracion = best_opt.iteracion + 1;
                 vecino.visto = true;
 
@@ -288,10 +288,10 @@ app.controller('mainCtrl', function($scope, $filter) {
                 opciones.push(vecino); //anadimos la nueva opcion
               }
               //Si el costo actual del vertice es mayor que el que nosotros ofrecemos
-              else if (vecino.visto && vecino.costo > (parseInt(best_opt.costo) + parseInt(conexion.costo))) {
+              else if (vecino.visto && vecino.costo > (parseFloat(best_opt.costo) + parseFloat(conexion.costo))) {
                 //alert("Sobrescrito!!: " + JSON.stringify(vecino));
-                vecino.padre = parseInt(best_opt.id, 10);
-                vecino.costo = (parseInt(best_opt.costo, 10) + parseInt(conexion.costo, 10));
+                vecino.padre = parseFloat(best_opt.id, 10);
+                vecino.costo = (parseFloat(best_opt.costo, 10) + parseFloat(conexion.costo, 10));
                 vecino.iteracion = best_opt.iteracion + 1;
                 //actualizamos como se ve el vertice en el array que vamos a retornar
                 for (var j = 1; j < v.length; j++) {
